Arreglos pago de credito y se adicion pago de credito de terceros

+++ lib/modules/home/UserSessionInfo/data/models/user_session_info_model.dart
@@ -94,8 +94,7 @@ class UserInfoResponseModel extends UserInfoResponseEnttity {
         idcQuestion: json["idcQuestion"],
         cellPhoneNumber: json["cellPhoneNumber"],
         maximumAmount: json["maximumAmount"].toDouble(),
-        maximumElectronicWalletAmount: json["maximumElectronicWalletAmount"]
-            .toDouble(),
+        maximumElectronicWalletAmount: json["maximumElectronicWalletAmount"].toDouble(),
         email: json["email"],
         sendEmailNotification: json["sendEmailNotification"],
         identityCardNumber: json["identityCardNumber"],
@@ -185,7 +184,7 @@ class ListCodeCreditLineElementModel extends ListCodeCreditLineElementEntity {
       ListCodeCreditLineElementModel(
         idMoney: json["idMoney"],
         codMoney: json["codMoney"],
-        idOperationEntity: json["idOperationEntity"].toDouble(),
+        idOperationEntity: json["idOperationEntity"],
         operationCode: json["operationCode"],
         availableAmount: json["availableAmount"].toDouble(),
         idOffice: json["idOffice"],
